# public/openapi.yaml
openapi: 3.1.0
info:
  title: Yamada Engei Works GPT API
  description: API for Custom GPTs to interact with the Yamada Engei Works farm management system.
  version: 1.0.0
servers:
  - url: https://yamada-engei-works.vercel.app/api/gpt
    description: Production Server
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
security:
  - ApiKeyAuth: []
paths:
...
  /analysis:
    get:
      summary: Get Past Data Analysis
      description: Retrieves recent work records and past crop cycles for analysis or looking back at previous works.
      operationId: getAnalysis
      parameters:
        - name: greenhouseName
          in: query
          description: Name of the greenhouse to filter by (e.g., "1号ハウス")
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
...